rs6000: Remove the ancient mfcr peepholes
These peepholes replace two mfcr;mask sequences by one mfcr;mask;mask
sequence.  On modern cpus, the original mfcr's were actually mfocrf,
but the new insn is an actual heavy-weight mfcr.  This is very bad
for performance.

The comment says there is a three cycle delay between two consecutive
mfcr insns.  This may have been true on rios, and it's true on 604,
but on 603, 750, 7400 it is just a single cycle (on 7450 it is two).

This is also a define_peephole, and we should get rid of those.

So this patch just removes the peepholes; the benefit is marginal at
best, and it so very hurts in other cases.


	* config/rs6000/rs6000.md (define_peepholes for two mfcr's): Delete.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@237062 138bc75d-0d04-0410-961f-82ee72b054a4